requirements.txt
google-genai>=1.0.0
fastapi[standard]>=0.115.0
uvicorn>=0.30.0
numpy>=1.26.0
httpx>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
datasets>=3.0.0
